Trait lending_project::traits::abacus_token::AbacusToken

pub trait AbacusToken: ContractEnv {
    type emitTransferEventsOutput: ImpliesReturn<Result<(), PSP22Error>>;
    type emitTransferEventAndDecreaseAllowanceOutput: ImpliesReturn<Result<(), PSP22Error>>;
    type getLendingPoolOutput: ImpliesReturn<AccountId>;

    // Required methods
    fn emit_transfer_events(
        &mut self,
        transfer_event_data: Vec<TransferEventData>
    ) -> Self::emitTransferEventsOutput;
    fn emit_transfer_event_and_decrease_allowance(
        &mut self,
        transfer_event_data: TransferEventData,
        from: AccountId,
        to: AccountId,
        decrease_allowance_by: Balance
    ) -> Self::emitTransferEventAndDecreaseAllowanceOutput;
    fn get_lending_pool(&self) -> Self::getLendingPoolOutput;
}

Required Associated Types

type emitTransferEventsOutput: ImpliesReturn<Result<(), PSP22Error>>

Output type of the respective trait message.

type emitTransferEventAndDecreaseAllowanceOutput: ImpliesReturn<Result<(), PSP22Error>>

Output type of the respective trait message.

type getLendingPoolOutput: ImpliesReturn<AccountId>

Output type of the respective trait message.

Required Methods

fn emit_transfer_events( &mut self, transfer_event_data: Vec<TransferEventData> ) -> Self::emitTransferEventsOutput

called by lending protocol whenever the state of user deposit or debt (aToken, vToken) is changed.

fn emit_transfer_event_and_decrease_allowance( &mut self, transfer_event_data: TransferEventData, from: AccountId, to: AccountId, decrease_allowance_by: Balance ) -> Self::emitTransferEventAndDecreaseAllowanceOutput

called whenever the state of user deposit or debt (aToken, vToken) is changed caller != on_behalf_of. case1: Alice has allowance on Bobs ATokens and Alice makes withdraw on behalf of Bob. case2: Alice has allowance on Bobs VTokens and Alice makes borrow on behalf of Bob.

fn get_lending_pool(&self) -> Self::getLendingPoolOutput

returns lending_pool AccountId