Module lending_project::traits::lending_pool

a trait implemented by core contract of the protocol. It comes with all the functionalities of Lending.

Macros

contract_ref

Generates a wrapper which can be used for interacting with the contract.

Structs

AssetRules

Defines rules on which asset can be borrowed and used as collateral.

ReserveAbacusTokens

Stores AccountIds of PSP22(PSP55) contracts that are wrappers of deposit(debt).

ReserveData

Contains most often used data of a reserve

ReserveIndexes

Stores data used to accumulate deposit and debt interest rates.

ReserveParameters

Stores parameters used to calculate current_debt_rate and current_deposit_rate. The Abax native stable tokens do not use it!

ReserveRestrictions

Stores restrictions made on the reserve

UserConfig

stores information about which asset is deposited, used as collatearl and borrowed by an user and which market_rule_id the user has chosen.

UserReserveData

stores data of user

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

DefaultEnvironment

The fundamental types of the default configuration.

LendingPoolError

Possible errors returned by LendingPool messages.

MathError

Traits

EmitBorrowEvents

EmitConfigureEvents

EmitDepositEvents

EmitFlashEvents

EmitLiquidateEvents

EmitMaintainEvents

EmitManageEvents

LendingPoolATokenInterface

Trait containing messages that are accessible to AToken - the PSP22 wrapeer of deposits.

LendingPoolBorrow

Trait containing messges that are used by borrowers.

LendingPoolDeposit

Trait containing messages that are used by depositors.

LendingPoolFlash

Trait containing flash_loan message. Used by flash loaners.

LendingPoolLiquidate

Trait containing liquidate message. Used by liquidators.

LendingPoolMaintain

Trait containing messages that are used to maintain inetrest accumulation. Used by maintainers.

LendingPoolManage

Trait containing AccessControl messages used to manage ‘LendingPool’ parameters. USed by managers.

LendingPoolVTokenInterface

Trait containing messages that are accessible to VToken - the PSP22 Wrapper of debts.

LendingPoolView

Trait containing non-mutable messages.

Type Definitions

AccountId

Aliases for types of the default environment

AssetId

type used to identify asset

Balance

Bitmap128

DecimalMultiplier

LendingPoolATokenInterfaceRef

LendingPoolBorrowRef

LendingPoolDepositRef

LendingPoolFlashRef

LendingPoolLiquidateRef

LendingPoolMaintainRef

LendingPoolManageRef

LendingPoolVTokenInterfaceRef

LendingPoolViewRef

MarketRule

type used to represent market rule

RuleId

type used to identigy rule

Timestamp