Trait lending_project::traits::lending_pool::LendingPoolManage

pub trait LendingPoolManage: ContractEnv {

    type setPriceFeedProviderOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type setFlashLoanFeeE6Output: ImpliesReturn<Result<(), LendingPoolError>>;
    type setBlockTimestampProviderOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type registerAssetOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type registerStablecoinOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type setReserveIsActiveOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type setReserveIsFreezedOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type setReserveParametersOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type setReserveRestrictionsOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type setStablecoinDebtRateE24Output: ImpliesReturn<Result<(), LendingPoolError>>;
    type addMarketRuleOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type modifyAssetRuleOutput: ImpliesReturn<Result<(), LendingPoolError>>;
    type takeProtocolIncomeOutput: ImpliesReturn<Result<Vec<(AccountId, i128)>, LendingPoolError>>;

    // Required methods
    fn set_price_feed_provider(
        &mut self,
        price_feed_provider: AccountId
    ) -> Self::setPriceFeedProviderOutput;
    fn set_flash_loan_fee_e6(
        &mut self,
        flash_loan_fee_e6: u128
    ) -> Self::setFlashLoanFeeE6Output;
    fn set_block_timestamp_provider(
        &mut self,
        provider_address: AccountId
    ) -> Self::setBlockTimestampProviderOutput;
    fn register_asset(
        &mut self,
        asset: AccountId,
        a_token_code_hash: [u8; 32],
        v_token_code_hash: [u8; 32],
        name: String,
        symbol: String,
        decimals: u8,
        asset_rules: AssetRules,
        maximal_total_deposit: Option<Balance>,
        maximal_total_debt: Option<Balance>,
        minimal_collateral: Balance,
        minimal_debt: Balance,
        income_for_suppliers_part_e6: u128,
        interest_rate_model: [u128; 7]
    ) -> Self::registerAssetOutput;
    fn register_stablecoin(
        &mut self,
        asset: AccountId,
        a_token_code_hash: [u8; 32],
        v_token_code_hash: [u8; 32],
        name: String,
        symbol: String,
        decimals: u8,
        asset_rules: AssetRules,
        maximal_total_deposit: Option<Balance>,
        maximal_total_debt: Option<Balance>,
        minimal_collateral: Balance,
        minimal_debt: Balance
    ) -> Self::registerStablecoinOutput;
    fn set_reserve_is_active(
        &mut self,
        asset: AccountId,
        active: bool
    ) -> Self::setReserveIsActiveOutput;
    fn set_reserve_is_freezed(
        &mut self,
        asset: AccountId,
        freeze: bool
    ) -> Self::setReserveIsFreezedOutput;
    fn set_reserve_parameters(
        &mut self,
        asset: AccountId,
        interest_rate_model: [u128; 7],
        income_for_suppliers_part_e6: u128
    ) -> Self::setReserveParametersOutput;
    fn set_reserve_restrictions(
        &mut self,
        asset: AccountId,
        maximal_total_deposit: Option<Balance>,
        maximal_total_debt: Option<Balance>,
        minimal_collateral: Balance,
        minimal_debt: Balance
    ) -> Self::setReserveRestrictionsOutput;
    fn set_stablecoin_debt_rate_e24(
        &mut self,
        asset: AccountId,
        debt_rate_e24: u128
    ) -> Self::setStablecoinDebtRateE24Output;
    fn add_market_rule(
        &mut self,
        market_rule: MarketRule
    ) -> Self::addMarketRuleOutput;
    fn modify_asset_rule(
        &mut self,
        market_rule_id: RuleId,
        asset: AccountId,
        asset_rules: AssetRules
    ) -> Self::modifyAssetRuleOutput;
    fn take_protocol_income(
        &mut self,
        assets: Option<Vec<AccountId>>,
        to: AccountId
    ) -> Self::takeProtocolIncomeOutput;

}

Trait containing AccessControl messages used to manage ‘LendingPool’ parameters. USed by managers.

Required Associated Types

type setPriceFeedProviderOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setFlashLoanFeeE6Output: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setBlockTimestampProviderOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type registerAssetOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type registerStablecoinOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setReserveIsActiveOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setReserveIsFreezedOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setReserveParametersOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setReserveRestrictionsOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type setStablecoinDebtRateE24Output: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type addMarketRuleOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type modifyAssetRuleOutput: ImpliesReturn<Result<(), LendingPoolError>>

Output type of the respective trait message.

type takeProtocolIncomeOutput: ImpliesReturn<Result<Vec<(AccountId, i128)>, LendingPoolError>>

Output type of the respective trait message.

Required Methods

fn set_price_feed_provider( &mut self, price_feed_provider: AccountId ) -> Self::setPriceFeedProviderOutput

Sets a price_feed_provider - a contract that implements PriceFeed and will be used to get prices from.

price_feed_provider AccountId (a.k.a. address) of a contract.

Errors

AccessControl::MisingRole returned if the caller is not a PARAMETERS_ADMIN.

fn set_flash_loan_fee_e6( &mut self, flash_loan_fee_e6: u128 ) -> Self::setFlashLoanFeeE6Output

Sets a flash_loan_fee_e6

flash_loan_fee_e6 fee to set 1_000_000 = 100% fee.

Errors

AccessControl::MisingRole returned if the caller is not a PARAMETERS_ADMIN.

fn set_block_timestamp_provider( &mut self, provider_address: AccountId ) -> Self::setBlockTimestampProviderOutput

only for tests sets provider of block_timestamp

fn register_asset( &mut self, asset: AccountId, a_token_code_hash: [u8; 32], v_token_code_hash: [u8; 32], name: String, symbol: String, decimals: u8, asset_rules: AssetRules, maximal_total_deposit: Option<Balance>, maximal_total_debt: Option<Balance>, minimal_collateral: Balance, minimal_debt: Balance, income_for_suppliers_part_e6: u128, interest_rate_model: [u128; 7] ) -> Self::registerAssetOutput

Registers new asset in the LendingPool’s storage and instaniates ‘AToken’ and ‘VToken’ for the reserve.

asset - AccountId of the registered asset

a_token_code_hash - code hash that will be used to initialize AToken

v_token_code_hash - code hash that will be used to initialize vToken

name - name of the asset. It will be used to create names for AToken and VToken.

symbol - symbol of the asset. It will be used to create sumbol for AToken and VToken.

decimals - a decimal denominator of an asset (number already multiplied by 10^N where N is number of decimals)

asset_rules' - asset`’s AssetRules that will be used in default market rule (id = 0).

maximal_total_deposit - maximal allowed total deposit. None for uncapped.

maximal_total_debt - maximal allowed total debt. None for uncapped.

minimal_collateral - the required minimal deposit of the asset by user to turn asset to be collateral.

minimal_debt - the minimal possible debt that can be taken by user.

interest_rate_model - targetted debt rates at utilization of 68%, 84%, 92%, 96%, 98%, 99% 100%

income_for_suppliers_part_e6 - indicates which part of an income should suppliers be paid - in E6 notation (multiplied by 10^6)

Errors

AccessControl::MisingRole returned if the caller is not a ASSET_LISTING_ADMIN.

AlreadyRegistered returned if asset was already registered.

InvalidAssetRule returned if asset rule is invalid.

fn register_stablecoin( &mut self, asset: AccountId, a_token_code_hash: [u8; 32], v_token_code_hash: [u8; 32], name: String, symbol: String, decimals: u8, asset_rules: AssetRules, maximal_total_deposit: Option<Balance>, maximal_total_debt: Option<Balance>, minimal_collateral: Balance, minimal_debt: Balance ) -> Self::registerStablecoinOutput

Registers new Abax native Stable asset in the LendingPool’s storage and instaniates ‘AToken’ and ‘VToken’ for the reserve.

asset - AccountId of the registered asset

a_token_code_hash - code hash that will be used to initialize AToken

v_token_code_hash - code hash that will be used to initialize vToken

name - name of the asset. It will be used to create names for AToken and VToken.

symbol - symbol of the asset. It will be used to create sumbol for AToken and VToken.

decimals - a decimal denominator of an asset (number already multiplied by 10^N where N is number of decimals)

asset_rules' - asset`’s AssetRules that will be used in default market rule (id = 0).

maximal_total_deposit - maximal allowed total deposit. None for uncapped.

maximal_total_debt - maximal allowed total debt. None for uncapped.

minimal_collateral - the required minimal deposit of the asset by user to turn asset to be collateral.

minimal_debt - the minimal possible debt that can be taken by user.

Errors

AccessControl::MisingRole returned if the caller is not a ASSET_LISTING_ADMIN.

AlreadyRegistered returned if asset was already registered.

InvalidAssetRule returned if asset rule is invalid.

fn set_reserve_is_active( &mut self, asset: AccountId, active: bool ) -> Self::setReserveIsActiveOutput

activates or disactivates reserve

active - true if reserve should be activated. False if reserve should be disactivated.

Errors

AccessControl::MisingRole returned if the caller is not a EMERGENCY_ADMIN.

AlreadySet returned if trying to set already set state.

fn set_reserve_is_freezed( &mut self, asset: AccountId, freeze: bool ) -> Self::setReserveIsFreezedOutput

Freezes or unfreezes reserve

freeze - true if reserve should be freezed. False if reserve should be unfreezed.

Errors

AccessControl::MisingRole returned if the caller is not a EMERGENCY_ADMIN.

AlreadySet returned if trying to set already set state.

fn set_reserve_parameters( &mut self, asset: AccountId, interest_rate_model: [u128; 7], income_for_suppliers_part_e6: u128 ) -> Self::setReserveParametersOutput

modifies ReserveParameters in the LendingPool’s storage

asset - AccountId of the registered asset

interest_rate_model - targetted debt rates at utilization of 50%, 60%, 70%, 80%, 90%, 95% 100%

income_for_suppliers_part_e6 - indicates which part of an income should suppliers be paid - in E6 notation (multiplied by 10^6)

Errors

AccessControl::MisingRole returned if the caller is not a PARAMETERS_ADMIN.

fn set_reserve_restrictions( &mut self, asset: AccountId, maximal_total_deposit: Option<Balance>, maximal_total_debt: Option<Balance>, minimal_collateral: Balance, minimal_debt: Balance ) -> Self::setReserveRestrictionsOutput

modifies ReserveRestricion in the LendingPool’s storage

asset - AccountId of the registered asset

maximal_total_deposit - maximal allowed total deposit, If exceeded no more deposits are accepted. None for uncapped total deposit.

maximal_total_debt - maximal allowed total debt, If exceeded no more borrows are accepted. None for uncapped total debt.

minimal_collateral - the required minimal deposit of the asset by user to turn asset to be collateral.

minimal_debt - the minimal possible debt that can be taken by user.

Errors

AccessControl::MisingRole returned if the caller is not a PARAMETERS_ADMIN.

fn set_stablecoin_debt_rate_e24( &mut self, asset: AccountId, debt_rate_e24: u128 ) -> Self::setStablecoinDebtRateE24Output

modifies the stablecoin debt rate

asset - AccountId of the registered stable asset

debt_rate_e24 - new debt rate

Errors

AccessControl::MisingRole returned if the caller is not a STABLECOIN_RATE_ADMIN.

AssetIsNotProtocolStablecoin returned if asset is not abax native stablecoin.

fn add_market_rule( &mut self, market_rule: MarketRule ) -> Self::addMarketRuleOutput

adds new market rule at next martket rule id

market_rule - list of asset rules for that market rule

Errors

AccessControl::MisingRole returned if the caller is not a STABLECOIN_RATE_ADMIN.

InvalidAssetRule returned if the market_rule contains invalid AssetRule.

fn modify_asset_rule( &mut self, market_rule_id: RuleId, asset: AccountId, asset_rules: AssetRules ) -> Self::modifyAssetRuleOutput

modifies asset_rules of a given asset in the market rule identified by market_rule_id

market_rule_id - id of market rule which shuuld be modified

asset - AccountId of a asset which rules should be modified

asset_rules' - asset's AssetRules that will be used in market rule with market_rule_id`.

Errors

AccessControl::MisingRole returned if the caller is not a STABLECOIN_RATE_ADMIN.

InvalidAssetRule returned if the asset_rules is invalid.

fn take_protocol_income( &mut self, assets: Option<Vec<AccountId>>, to: AccountId ) -> Self::takeProtocolIncomeOutput

collects income generated by the protocol

assets - vector of assets that income should be taken. If None takes income from all registered assets.

to - account which will receive income.

Errors

AccessControl::MisingRole returned if the caller is not a TREASURY.